refactor(fretboard): extract svg background
Move FretboardBackground and useWoodGrainTexture out of FretboardSVG.tsx
into their own dedicated files.
